requirements.txt
pandas==2.1.4
openpyxl==3.1.2
reportlab==4.0.8
python-docx==1.0.1
pdfplumber==0.10.3
pdf2docx==0.5.6
Pillow==10.1.0
PyQt6==6.6.1
PyInstaller==6.3.0 